% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pythonInterface.R
\docType{class}
\name{from_Python-class}
\alias{from_Python-class}
\title{Class for General Python Class Objects}
\description{
The Python side of the interface will return a general object from a Python class as an R
object of class "from_Python".  Its Python fields (converted to R objects) can be accessed by the \code{$}
operator.
}
\section{Slots}{

\describe{
\item{\code{serverClass}}{the Python type.}

\item{\code{module}}{the Python module, or ""}

\item{\code{fields}}{the converted versioin of the Python fields; these are accessed by the \code{$} operator.}
}}

