% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonInterface.R
\name{okPython}
\alias{okPython}
\title{Check for a Valid Python for Interface}
\usage{
okPython(verbose = FALSE)
}
\arguments{
\item{verbose}{Should a message with the cause of a failure be reported?  Default \code{FALSE}.}
}
\description{
The function returns true or false according to whether a Python interface can be established.  This
will fail if no Python exists, if it is incompatible with this version of XRPython (e.g., 32 vs 64 bits
in Windows), or if for some reason it can't evaluate a trivial expression correctly.  Warnings are printed
but ignored.
}
