% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/classStructure.R
\docType{class}
\name{PythonClassDef-class}
\alias{PythonClassDef}
\alias{PythonClassDef-class}
\title{##Class and Generator for Python Class Description from Python Metadata
##
##Extends the \code{"ServerClassDef"} class in the XR package.
##
##}
\description{
##Class and Generator for Python Class Description from Python Metadata
##
##Extends the \code{"ServerClassDef"} class in the XR package.
##
##
}
\section{Fields}{

\describe{
\item{\code{className}}{the name of the Python class
##}

\item{\code{module}}{the name of the Python module}
}}

