% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classStructure.R
\docType{class}
\name{PythonFunction-class}
\alias{PythonFunction}
\alias{PythonFunction-class}
\title{Proxy Objects in R for Python Functions}
\description{
A class and generator function for proxies in R for Python functions.
}
\details{
An object from this class is an R function that is a proxy for a function in Python. Calls to the R function evaluate
a call to the Python function.  The arguments in the call are converted to equivalent Python objects;
these typically include proxy objects for results previously computed through the XRPython interface.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the server language function}

\item{\code{module}}{the name of the module, if that needs to be imported}

\item{\code{evaluatorClass}}{the class for the evaluator, by default and usually, \code{\link{PythonInterface}}}

\item{\code{serverDoc}}{the docstring from Python, if any.}

\item{\code{serverArgs}}{the Python argument names (not currently used).}
}}

