% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pythonInterface.R
\docType{class}
\name{PythonObject-class}
\alias{PythonObject}
\alias{PythonObject-class}
\title{Proxy Objects in R for Python Objects}
\description{
This is a class for all proxy objects from a Python class with an R proxy class definition.
Objects will normally be from a subclass of this class, for the specific Python class.
}
\details{
Proxy objects returned from the Python interface will be promoted to objects
from a specific R proxy class for their Python class, if such a class has been defined.
}

