% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pythonInterface.R
\name{RPython}
\alias{RPython}
\title{An Evaluator for the Python Interface.}
\usage{
RPython(...)
}
\description{
Returns an evaluator for the Python interface.  Starts one on the first call, or if arguments are provided;
providing argument \code{.makeNew = TRUE} will force a new evaluator.  Otherwise, the current evaluator is
returned.
}
\details{
See \code{\link{PythonInterface}} for details of the evaluator.
}

