% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonInterface.R
\name{setupStep}
\alias{setupStep}
\title{The Setup Step}
\description{
The file \code{"setup.R"} in the \code{tools} directory is designed to create an explicit definiion of proxy
classes for the \code{"list"} and \code{"dict"} types in Python.  The file would be run through
\code{XR::packageSetup()} when creating or modifying these classes in the \code{XRPython} package.  It provides
a useful example for the general task of creating an explicit, written-out version of a proxy class.
}
\details{
The setup step generates a file \code{"R/pythonProxyClasses.R"} in the source directory for the package.  The
setup step needs to be run twice, first to generate the R code in that file, and again to use the
\code{roxygen2} package to generate documentation.

For the first round, the package needs to be installed with an empty version of the file (the file has to exist
because the package uses a \code{Collate:} directive that mentions it.  Running \code{packageSetup()} this time
defines the proxy classes and dumps them (with some extra stuff) to the target file.  and  adds a line to the
\code{NAMESPACE} to export both classes. (If we were willing to let \code{roxygenize()} create the namespace
directives this would be automatic, but I'm not willing.)

Now the package needs to be installed again, this time with the proxy classes, The second pass of the setup file
runs \code{roxygenize()}.  Finally, as usual with \code{roxygenize()},
the package has to be installed one more time to generate the actual documentation.
}
