\name{XRPython-package}
\alias{XRPython-package}
\alias{XRPython}
\docType{package}
\title{
Functional and Object-based Interface from R to Python
}
\description{
An Interface from R to Python, extending the 'reticulate' package by
using the general 'XR' package mechanism, supporting R proxies for
Python functions and classes.  Python results are kept as proxies in a
session dictionary, with objects, functions, fields and methods directly
available from R.
}
\author{
John M. Chambers
}
\references{
 Chambers, John M.
\emph{Extending R} Chapman & Hall/CRC 2016.
}
\keyword{ package }
