% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pythonInterface.R
\name{functions}
\alias{functions}
\alias{pythonAddToPath}
\alias{pythonGet}
\alias{pythonSend}
\title{Function Versions of Methods for Python Interface evaluators.}
\usage{
pythonSend(object, evaluator = XR::getInterface(.PythonInterfaceClass))

pythonGet(object, evaluator = XR::getInterface(.PythonInterfaceClass))

pythonAddToPath(..., evaluator = XR::getInterface(.PythonInterfaceClass))
}
\arguments{
\item{object}{an R object, to be sent to Python (\code{pythonSend()}) or a proxy object for
the Python object to be converted (\code{pythonGet()}).}

\item{evaluator}{The evaluator object to use.  By default, and usually, the current evaluator
is used, and one is started if none has been.}

\item{...}{arguments \code{directory}, \code{package} and \code{pos}.}
}
\description{
Function Versions of Methods for Python Interface evaluators.
}
\section{Functions}{
\itemize{
\item \code{pythonSend}: sends the \code{object} to Python, converting it via methods for
\code{\link[XR]{asServerObject}} and returns a proxy for the converted object.

\item \code{pythonGet}: converts the proxy object that is its argument to an \R{} object.

\item \code{pythonAddToPath}: adds the directory specified to the search path for Python objects.
}}

