%% NOT roxygen It has a bug for inherited reference classes.  Until
%% fix is installed cant use it here.
\docType{class}
\name{PythonClassDef-class}
\alias{PythonClassDef}
\alias{PythonClassDef-class}
\title{Class and Generator for Python Class Description from Python Metadata}
\description{
A description of a Python class, obtained from the Python class
definition object.
 This class extends the \code{"ServerClassDef"} class in the XR package.
}
\section{Fields}{

\describe{
\item{\code{className}}{the name of the Python class
##}

\item{\code{module}}{the name of the Python module}
}}

