% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classStructure.R
\docType{class}
\name{PythonClassDef-class}
\alias{PythonClassDef}
\alias{PythonClassDef-class}
\title{Class and Generator for Python Class Description from Python Metadata}
\arguments{
\item{evaluator}{the interface evaluator; by default and usually, the current Python evaluator.}
}
\description{
Creates a class definition object consistent with the XR structure.
See \code{\link{setPythonClass}} for details and arguments to the generator for the class.
}

