\name{XRPython-package}
\alias{XRPython-package}
\alias{XRPython}
\docType{package}
\title{
Functional and Object-based Interface from R to Python
}
\description{
An Interface from R to Python using, in a very preliminary way, the
principles of functional and object-based interface design presented
in the reference.
}
\author{
John M. Chambers
}
\references{
 Chambers, John M. (book in progress)
\emph{Extending R}
}
\keyword{ package }
